Change int to int64 for MaxReposSizeLimit

+++ modules/auth/admin.go
@@ -41,7 +41,7 @@ type AdminEditUserForm struct {
 	AllowImportLocal        bool
 	AllowCreateOrganization bool
 	ProhibitLogin           bool
-	MaxTotalReposSizeLimit	int
+	MaxTotalReposSizeLimit	int64
 }
 
 // Validate validates form fields
